Ripped out ReachAnnotation and had Reach adopt MKAnnotation directly which returned the speed and ability to load all the reaches on the map
